Refactor day 10 solution

+++ src/day_template/mod.rs
@@ -4,7 +4,7 @@ const DAY: u32 = 0;
 
 fn get_input() -> Vec<String> {
     use std::io::BufRead;
-    download_day((DAY) as u32, "input").unwrap();
+    download_day(DAY, "input").unwrap();
 
     let file = std::fs::File::open(format!("input/input{DAY}.txt")).unwrap();
     let reader = std::io::BufReader::new(&file);
